Add configurable copyable markers logic
