+++ packages/ckeditor5-clipboard/src/clipboardpipeline.ts
@@ -32,8 +32,7 @@ import ClipboardObserver, {
 import plainTextToHtml from './utils/plaintexttohtml.js';
 import normalizeClipboardHtml from './utils/normalizeclipboarddata.js';
 import viewToPlainText from './utils/viewtoplaintext.js';
-
-import { collectAndRemoveFakeMarkers, insertAndCollectFakeMarkers } from './utils/insertAndCollectFakeMarkers.js';
+import ClipboardMarkersUtils from './clipboardmarkersutils.js';
 
 // Input pipeline events overview:
 //
@@ -152,6 +151,13 @@ export default class ClipboardPipeline extends Plugin {
 	/**
 	 * @inheritDoc
 	 */
+	public static get requires() {
+		return [ ClipboardMarkersUtils ] as const;
+	}
+
+	/**
+	 * @inheritDoc
+	 */
 	public init(): void {
 		const editor = this.editor;
 		const view = editor.editing.view;
@@ -172,17 +178,13 @@ export default class ClipboardPipeline extends Plugin {
 		selection: Selection | DocumentSelection,
 		method: 'copy' | 'cut' | 'dragstart'
 	): void {
-		this.editor.model.change( writer => {
-			const insertedFakeMarkersElements = insertAndCollectFakeMarkers( writer, selection );
-			const documentFragment = this.editor.model.getSelectedContent( selection );
+		const clipboardMarkersUtils: ClipboardMarkersUtils = this.editor.plugins.get( 'ClipboardMarkersUtils' );
+		const documentFragment = clipboardMarkersUtils._copySelectedFragmentWithMarkers( selection );
 
-			collectAndRemoveFakeMarkers( writer, documentFragment, insertedFakeMarkersElements );
-
-			this.fire<ClipboardOutputTransformationEvent>( 'outputTransformation', {
-				dataTransfer,
-				content: documentFragment,
-				method
-			} );
+		this.fire<ClipboardOutputTransformationEvent>( 'outputTransformation', {
+			dataTransfer,
+			content: documentFragment,
+			method
 		} );
 	}
