+++ packages/ckeditor5-clipboard/src/index.ts
@@ -19,6 +19,8 @@ export {
 	type ViewDocumentClipboardOutputEvent
 } from './clipboardpipeline.js';
 
+export { default as ClipboardMarkersUtils } from './clipboardmarkersutils.js';
+
 export type {
 	ClipboardEventData
 } from './clipboardobserver.js';
@@ -27,8 +29,6 @@ export { default as DragDrop } from './dragdrop.js';
 export { default as PastePlainText } from './pasteplaintext.js';
 export { default as DragDropTarget } from './dragdroptarget.js';
 export { default as DragDropBlockToolbar } from './dragdropblocktoolbar.js';
-
-export { collectAndRemoveFakeMarkers, insertAndCollectFakeMarkers } from './utils/insertAndCollectFakeMarkers.js';
 
 export type {
 	ViewDocumentClipboardInputEvent,
